+++ src/components/interactive-map.js
@@ -3,9 +3,12 @@ import PropTypes from 'prop-types';
 import autobind from '../utils/autobind';
 
 import StaticMap from './static-map';
-import MapControls from './map-controls';
 import MapState, {MAPBOX_MAX_PITCH, MAPBOX_MAX_ZOOM} from '../utils/map-state';
 import {PerspectiveMercatorViewport} from 'viewport-mercator-project';
+
+import EventManager from '../utils/event-manager/event-manager';
+import MapControls from '../utils/map-controls';
+import config from '../config';
 
 const propTypes = Object.assign({}, StaticMap.propTypes, {
   // Additional props on top of StaticMap
@@ -38,15 +41,33 @@ const propTypes = Object.assign({}, StaticMap.propTypes, {
    */
   onChangeViewport: PropTypes.func,
 
+  /** Enables perspective control event handling */
+  perspectiveEnabled: PropTypes.bool,
+
+  /**
+    * Is the component currently being dragged. This is used to show/hide the
+    * drag cursor. Also used as an optimization in some overlays by preventing
+    * rendering while dragging.
+    */
+  isHovering: PropTypes.bool,
+  isDragging: PropTypes.bool,
+
   /** Advanced features */
   // Contraints for displaying the map. If not met, then the map is hidden.
   displayConstraints: PropTypes.object.isRequired,
-  // A React component class definition to replace the default map controls
-  ControllerClass: PropTypes.func
+  // A map control instance to replace the default map controls
+  // The object must expose one property: `events` as an array of subscribed
+  // event names; and two methods: `setState(state)` and `handle(event)`
+  mapControls: PropTypes.shape({
+    events: PropTypes.arrayOf(PropTypes.string),
+    setState: PropTypes.func,
+    handle: PropTypes.func
+  })
 });
 
 const defaultProps = Object.assign({}, StaticMap.defaultProps, {
   onChangeViewport: null,
+  perspectiveEnabled: false,
 
   /** Viewport constraints */
   maxZoom: MAPBOX_MAX_ZOOM,
@@ -58,7 +79,8 @@ const defaultProps = Object.assign({}, StaticMap.defaultProps, {
     maxZoom: MAPBOX_MAX_ZOOM,
     maxPitch: MAPBOX_MAX_PITCH
   },
-  ControllerClass: MapControls
+
+  mapControls: new MapControls()
 });
 
 export default class InteractiveMap extends PureComponent {
@@ -78,9 +100,56 @@ export default class InteractiveMap extends PureComponent {
     };
   }
 
+  componentDidMount() {
+    // Register event handlers
+    const {eventCanvas} = this.refs;
+    const {mapControls} = this.props;
+
+    this._eventManager = new EventManager(eventCanvas);
+
+    mapControls.events.forEach(event => this._eventManager.on(event, this._handleEvent));
+  }
+
+  componentWillUnmount() {
+    if (this._eventManager) {
+      // Must destroy because hammer adds event listeners to window
+      this._eventManager.destroy();
+    }
+  }
+
+  _handleEvent(event) {
+    const {mapControls} = this.props;
+    // MapControls only extracts the states that it recognizes.
+    // This allows custom map controls to add new states and callbacks.
+    mapControls.setState(Object.assign({}, this.props, {
+      mapState: new MapState(this.props)
+    }));
+
+    return mapControls.handle(event);
+  }
+
   // TODO - Remove once Viewport alternative is good enough
   _getMap() {
     return this._map._getMap();
+  }
+
+  // Calculate a cursor style to show that we are in "dragging state"
+  _getCursor() {
+    const isInteractive =
+      this.props.onChangeViewport ||
+      this.props.onClickFeature ||
+      this.props.onHoverFeatures;
+
+    if (!isInteractive) {
+      return 'inherit';
+    }
+    if (this.props.isDragging) {
+      return config.CURSOR.GRABBING;
+    }
+    if (this.props.isHovering) {
+      return config.CURSOR.POINTER;
+    }
+    return config.CURSOR.GRAB;
   }
 
   // Checks a displayConstraints object to see if the map should be displayed
@@ -104,7 +173,7 @@ export default class InteractiveMap extends PureComponent {
   }
 
   render() {
-    const {width, height, ControllerClass} = this.props;
+    const {width, height} = this.props;
     const mapVisible = this.checkDisplayConstraints(this.props);
     const visibility = mapVisible ? 'visible' : 'hidden';
     const overlayContainerStyle = {
@@ -116,12 +185,19 @@ export default class InteractiveMap extends PureComponent {
       overflow: 'hidden'
     };
 
+    const eventCanvasStyle = {
+      width,
+      height,
+      position: 'relative',
+      cursor: this._getCursor()
+    };
+
     return (
-      createElement(ControllerClass, Object.assign({}, this.props, {
+      createElement('div', {
         key: 'map-controls',
-        style: {position: 'relative'},
-        mapState: new MapState(this.props)
-      }), [
+        ref: 'eventCanvas',
+        style: eventCanvasStyle
+      }, [
         createElement(StaticMap, Object.assign({}, this.props, {
           key: 'map-static',
           style: {visibility},